Moved section header to proper folder
